pyproject.toml
[tool.poetry]
name = "python_generator"
version = "0.1.0"
description = ""
authors = ["Patrick Rodrigues <patrick.rodrigues@olist.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
